requirements.txt
flask==3.0.2
request==2.31.0
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
pydantic==2.6.2
python-dotenv==1.0.1